Transferer.py: Backport production fixes into HEAD (xtrabackup in path)
Backport 361a1ef into the transferpy repo:

The paths to xtrabackup and mbstream were hard-coded to the binaries
on wmf-mariadb101 (mariabackup and mbstream).

Use PATH to get xtrabackup and mbstream (puppet, or preferably, the
package will take care of the install), so this tool can be used
with the general xtrabackup or maria-backup packages, not just on
WMF installations.

This will also allow to start backing up 10.4 hosts or later
versions, not only 10.1 ones.

Bug: T250666
Change-Id: Ie2aa349a37cc74a195cee4934666f542bc6ab382