transferpy: Remove wmfmariadbpy package
transferpy was a subpackage of the wmfmariadbpy project. For easiness
in development and use, wmfmariadbpy project divided into two separate
the project, namely transferpy and wmfmariadbpy. While separation we
agreed-upon replicating the wmfmariadbpy on both repositories and
remove the unnecessary files later in a patch like this so that
the history project remains intact.

So, for the above purpose, this patch removes the wmfmariadbpy
related files and configurations from the transferpy repository.

Bug: T248256
Change-Id: Ia1414c29807eb5f3fe450731a519b6e9457f3932