Firewall.py: Solve auto port detection concurrency issue
transferpy has the capability to detect a free port automatically
for netcat to listen. Simultaneous starting of two transferpy
instances for the same destination machine may create the possibility
of both instances detecting the same port as the free port
(called as race condition).

This patch solves this problem by introducing a reservation mechanism
leveraging the Linux atomicity in making directories. This patch
updates the Firewall.py so that, after detecting a free port,
transferpy try reserving it using mkdir. If mkdir fails because of
some other transferpy instance booked this port in between, it will
find another open port and make the reservation try again. This try
continues until we find an actual free port or max number of
attempts reached.

Bug: T256450
Change-Id: I5782057603f25e119b3bcece8d5dd41b98642537