Make transferpy configurable using a configuration file
Currently, transferpy takes all the user arguments from the command line.
This patch makes it configurable using a configuration file. The first
priority goes to command-line arguments, second to the configuration
file, and third to default values in the code.

This patch uses the configparser python module to read the
configuration file.

Bug: T257602, T254979
Change-Id: I09a7cf671d39e99fbd532f8ed5aa9ffc323a964b