Transferer.py: Add proper cleanup
The transferpy creates temporary files for the port-locking and checksum
creation. Interruptions in the middle of a transfer may leave these files
undeleted. This patch adds a proper cleanup method called
clean_all_temps to the transferpy. This method at the `finally` option
resolves the undeletion problem.

This patch introduces two new exceptions: FirewallError and TempDeletionError.

Bug: T256450
Change-Id: I8a24757ae25ba04502fab3d2499a031de6a8c299