Firewall.py: Save the target port after reservation
Currently, the Firewall only saves the target_host. This
patch makes the Firewall object to keep the opened port
of the target_host also. This patch also changes the close
function signature because now the close function does not
need a target_port argument since it is already saved
inside the Firewall.

Bug: T257601
Change-Id: Icec2306e1bae95e692157882877942300777bc25